test: extract OAuth integration tests into their own file
Move token rotation, client sweep, and reuse detection tests from
server-integration.test.ts into server-oauth-integration.test.ts.
Revert the knip.json workaround and restore the original Revoked
tokens bullet wording in ARCHITECTURE.md.
