+++ knip.json
@@ -1,6 +1,5 @@
 {
   "entry": ["cli/src/bin.ts"],
   "project": ["src/**/*.ts", "cli/src/**/*.ts", "scripts/**/*.ts"],
-  "ignoreBinaries": ["optipng", "knip", "markdownlint-cli2"],
-  "ignoreDependencies": ["markdownlint-cli2"]
+  "ignoreBinaries": ["optipng"]
 }
